add func insert and findbyid Movie

+++ app/src/main/java/com/example/movie_app_compose/model/Detail.kt
@@ -76,7 +76,7 @@ data class Detail(
     val video: Boolean?,
 
     @field:SerializedName("vote_average")
-    val vote_average: Double?,
+    val vote_average: Float?,
 
     @field:SerializedName("vote_count")
     val vote_count: Int?
